ECS Fargate example
Worked through an example to get this code up and running for our preview
environment. Rhagolwg for the win!

This is a poor example of how to fork but code was originally taken from here:
https://github.com/duduribeiro/terraform_ecs_fargate_example

This currently `plans` green but uses original rails secrets and hasn't been
modified for polygon8's benzaiten or hotei parts yet. It will probably not
apply cleanly.